requirements.txt
fastapi
uvicorn
msgspec
requests
eth-abi
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
python-multipart

